Problems with NUnit Testing
While using NUnit, I kept getting errors for the test:

First error I got was telling me the start tag of 'orders' didn't match the end tag of 'title', which most likely meant the XML was malformed.  After using the W3C validator to check it, I changed the XmlTest1 string to a literal string and got rid of all the quotes and whatnot and that seemed to help.

Second error was regarding the date.  In the assert for the date, the value from the XML was being formatted with toShortDateString() and was being compared to a string of "01/01/2009".  When the assert was done, toShortDateString() changed the date to "1/1/2009" due to the US culture and so it threw an error that said it was expecting 10 digits but only got 8.  So, I converted the "01/01/2009" string to a date and used toShortDateString() on it, which then passed the assertion.